refactor(Prediction): Move JPA related stuff to `db` folder
